basic graphql server
