+++ server/README.md
@@ -1,0 +1,12 @@
+reference for graphql server: https://oliverjumpertz.com/how-to-build-a-powerful-graphql-api-with-rust/
+
+run `cargo run` and launch `localhost:8080`.
+
+Execute the query:
+```
+query Query {
+    hello
+}
+```
+
+Also try `localhost:8080/health`.
