+++ server/src/model/mod.rs
@@ -1,0 +1,11 @@
+use async_graphql::{Context, Object, Schema};
+use async_graphql::{EmptyMutation, EmptySubscription};
+pub(crate) type ServiceSchema = Schema<QueryRoot, EmptyMutation, EmptySubscription>;
+
+pub(crate) struct QueryRoot;
+#[Object]
+impl QueryRoot {
+    async fn hello(&self, _ctx: &Context<'_>) -> &'static str {
+        "Hello world"
+    }
+}
