+++ server/src/routes/mod.rs
@@ -1,0 +1,33 @@
+use crate::model::ServiceSchema;
+use async_graphql::http::{playground_source, GraphQLPlaygroundConfig};
+use async_graphql_axum::{GraphQLRequest, GraphQLResponse};
+use axum::{
+    extract::Extension,
+    http::StatusCode,
+    response::{Html, IntoResponse},
+    Json,
+};
+use serde::Serialize;
+
+#[derive(Serialize)]
+struct Health {
+    healthy: bool,
+}
+
+pub(crate) async fn health() -> impl IntoResponse {
+    let health = Health { healthy: true };
+    (StatusCode::OK, Json(health))
+}
+
+pub(crate) async fn graphql_playground() -> impl IntoResponse {
+    Html(playground_source(
+        GraphQLPlaygroundConfig::new("/").subscription_endpoint("/ws"),
+    ))
+}
+
+pub(crate) async fn graphql_handler(
+    schema: Extension<ServiceSchema>,
+    req: GraphQLRequest,
+) -> GraphQLResponse {
+    schema.execute(req.into_inner()).await.into()
+}
